requirements.txt
Package             Version
------------------- -------
Django              2.1.1  
djangorestframework 3.8.2  
Pillow              5.3.0  
pip                 18.0   
pytz                2018.5 
setuptools          40.4.3 
wheel               0.32.0 

